more qreal=double assumptions fixed (yay arm)

+++ krita/plugins/paintops/libbrush/kis_auto_brush.cpp
@@ -62,8 +62,8 @@ KisAutoBrush::KisAutoBrush(KisMaskGenerator* as, qreal angle, qreal randomness, 
     d->density = density;
     d->idealThreadCountCached = QThread::idealThreadCount();
     setBrushType(MASK);
-    setWidth(qMax(1.0, d->shape->width()));
-    setHeight(qMax(1.0, d->shape->height()));
+    setWidth(qMax(qreal(1.0), d->shape->width()));
+    setHeight(qMax(qreal(1.0), d->shape->height()));
     setAngle(angle);
     QImage image = createBrushPreview();
     setImage(image);
